fixed floatfield on models, fixed revenue create view

+++ pointofsaleapi/models/revenue.py
@@ -3,7 +3,7 @@ from .order import Order
 
 class Revenue(models.Model):
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
-    total = models.IntegerField()
+    total = models.FloatField()
     payment_type = models.CharField(max_length=20)
-    tip = models.IntegerField()
+    tip = models.FloatField()
     order_type = models.CharField(max_length=20)
